Extract MultiParentCasperFinalizationSpec
