+++ casper/src/test/scala/coop/rchain/casper/MultiParentCasperRemainingTest.scala
@@ -368,90 +368,6 @@ class MultiParentCasperRemainingTest extends FlatSpec with Matchers with Inspect
     } yield result
   }
 
-  it should "increment last finalized block as appropriate in round robin" in effectTest {
-    val stake      = 10L
-    val equalBonds = validators.map(_ -> stake).toMap
-    val genesisWithEqualBonds =
-      buildGenesis(Seq.empty, equalBonds, 1L, Long.MaxValue, Faucet.noopFaucet, 0L)
-    for {
-      nodes       <- HashSetCasperTestNode.networkEff(validatorKeys.take(3), genesisWithEqualBonds)
-      deployDatas <- (0 to 7).toList.traverse(i => ConstructDeploy.basicDeployData[Effect](i))
-
-      createBlock1Result <- nodes(0).casperEff
-                             .deploy(deployDatas(0)) *> nodes(0).casperEff.createBlock
-      Created(block1) = createBlock1Result
-      _               <- nodes(0).casperEff.addBlock(block1, ignoreDoppelgangerCheck[Effect])
-      _               <- nodes(1).receive()
-      _               <- nodes(2).receive()
-
-      createBlock2Result <- nodes(1).casperEff
-                             .deploy(deployDatas(1)) *> nodes(1).casperEff.createBlock
-      Created(block2) = createBlock2Result
-      _               <- nodes(1).casperEff.addBlock(block2, ignoreDoppelgangerCheck[Effect])
-      _               <- nodes(0).receive()
-      _               <- nodes(2).receive()
-
-      createBlock3Result <- nodes(2).casperEff
-                             .deploy(deployDatas(2)) *> nodes(2).casperEff.createBlock
-      Created(block3) = createBlock3Result
-      _               <- nodes(2).casperEff.addBlock(block3, ignoreDoppelgangerCheck[Effect])
-      _               <- nodes(0).receive()
-      _               <- nodes(1).receive()
-
-      createBlock4Result <- nodes(0).casperEff
-                             .deploy(deployDatas(3)) *> nodes(0).casperEff.createBlock
-      Created(block4) = createBlock4Result
-      _               <- nodes(0).casperEff.addBlock(block4, ignoreDoppelgangerCheck[Effect])
-      _               <- nodes(1).receive()
-      _               <- nodes(2).receive()
-
-      createBlock5Result <- nodes(1).casperEff
-                             .deploy(deployDatas(4)) *> nodes(1).casperEff.createBlock
-      Created(block5) = createBlock5Result
-      _               <- nodes(1).casperEff.addBlock(block5, ignoreDoppelgangerCheck[Effect])
-      _               <- nodes(0).receive()
-      _               <- nodes(2).receive()
-
-      _     <- nodes(0).casperEff.lastFinalizedBlock shouldBeF block1
-      state <- nodes(0).casperState.read
-      _     = state.deployHistory.size should be(1)
-
-      createBlock6Result <- nodes(2).casperEff
-                             .deploy(deployDatas(5)) *> nodes(2).casperEff.createBlock
-      Created(block6) = createBlock6Result
-      _               <- nodes(2).casperEff.addBlock(block6, ignoreDoppelgangerCheck[Effect])
-      _               <- nodes(0).receive()
-      _               <- nodes(1).receive()
-
-      _     <- nodes(0).casperEff.lastFinalizedBlock shouldBeF block2
-      state <- nodes(0).casperState.read
-      _     = state.deployHistory.size should be(1)
-
-      createBlock7Result <- nodes(0).casperEff
-                             .deploy(deployDatas(6)) *> nodes(0).casperEff.createBlock
-      Created(block7) = createBlock7Result
-      _               <- nodes(0).casperEff.addBlock(block7, ignoreDoppelgangerCheck[Effect])
-      _               <- nodes(1).receive()
-      _               <- nodes(2).receive()
-
-      _ <- nodes(0).casperEff.lastFinalizedBlock shouldBeF block3
-      _ = state.deployHistory.size should be(1)
-
-      createBlock8Result <- nodes(1).casperEff
-                             .deploy(deployDatas(7)) *> nodes(1).casperEff.createBlock
-      Created(block8) = createBlock8Result
-      _               <- nodes(1).casperEff.addBlock(block8, ignoreDoppelgangerCheck[Effect])
-      _               <- nodes(0).receive()
-      _               <- nodes(2).receive()
-
-      _     <- nodes(0).casperEff.lastFinalizedBlock shouldBeF block4
-      state <- nodes(0).casperState.read
-      _     = state.deployHistory.size should be(1)
-
-      _ <- nodes.map(_.tearDown()).toList.sequence
-    } yield ()
-  }
-
   private def buildBlockWithInvalidJustification(
       nodes: IndexedSeq[HashSetCasperTestNode[Effect]],
       deploys: immutable.IndexedSeq[ProcessedDeploy],
